day2/lab 4/requirements.txt
langchain
langchain-groq
langchain-chroma
langchain-huggingface
langchain-community
chromadb
streamlit
python-dotenv
sentence-transformers
pypdf
tiktoken
